fixed xpm deallocation and switched from int to double for mlx_get_time (more precision)

+++ src/utils/mlx_utils.c
@@ -118,7 +118,7 @@ uint32_t mlx_rgba_to_mono(uint32_t color)
 
 //= Public =//
 
-int32_t mlx_get_time(void)
+double mlx_get_time(void)
 {
 	return (glfwGetTime());
 }
